Nuclear Reactor (#130)
* First Nuclear Rework

* Nuclear Reactor REI (WIP)

* Add Nuclear Reactor GUI (WIP)

* Nuclear Code Refactor

* Rework Button in Screen handler

* Refactor and balance Nuclear Fuel

* Refactor Nuclear Fuel

* Add Cryofluid

* Add Blast Furnace to Java

* Add Nuclear Fuel Rod Pipeline

* Add Nuclear Fuel Recycling

* Fuel Fully Work in the reactor

* Add Endstone Recipe

* Prepare for fluid in the reactor

* Refactor MIStorage and use transaction for fluid heating

* Fix Storage Impl

* MVP Nuclear Reactor

* Add Fluid to Nuclear Interaction in REI

* Fluid Neutron Product

* Add Thermal Properties in REI

* Fix Numerical issues

* Finish Rework Nuclear Reactor GUI

* Add Remaning Component for MVP

* New Recipe & Make Easy Mode The Only Recipes

* Add Nuclear Reactor Recipe

+++ src/main/java/aztech/modern_industrialization/pipes/fluid/FluidNetwork.java
@@ -67,7 +67,7 @@ public class FluidNetwork extends PipeNetwork {
                 loadedNodeCount++;
             }
         }
-        long networkCapacity = loadedNodeCount * nodeCapacity;
+        long networkCapacity = (long) loadedNodeCount * nodeCapacity;
         FluidVariant fluid = ((FluidNetworkData) data).fluid;
         // If the fluid is EMPTY we stop here
         if (fluid.isBlank()) {
